split functions
Разбил функции ( и пока что классы) на отдельные файлы
