+++ my-app/src/game.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import ReactDOM  from 'react-dom';
+import './index.css';
+import { Board } from './board';
+
+export class Game extends React.Component {
+    render() {
+        return (
+            <div className="game">
+                <div className="game-board">
+                    <Board />
+                </div>
+                <div className="game-info">
+                    <div>{/* status */}</div>
+                    <ol>{/* TODO */}</ol>
+                </div>
+            </div>
+        );
+    }
+}
